--- a/grug/app/src/traits.rs
+++ b/grug/app/src/traits.rs
@@ -1,8 +1,8 @@
 use {
     crate::{GasTracker, QuerierProvider, StorageProvider},
+    borsh::{BorshDeserialize, BorshSerialize},
     grug_types::{Batch, Context, Hash256, Storage},
     ics23::CommitmentProof,
-    serde::{de::DeserializeOwned, ser::Serialize},
 };
 
 // ------------------------------------ db -------------------------------------
@@ -42,7 +42,7 @@ pub trait Db {
     type StateCommitment: Storage + Clone + 'static;
 
     /// Type of the Merkle proof. The DB can choose any Merkle tree scheme.
-    type Proof: Serialize + DeserializeOwned;
+    type Proof: BorshSerialize + BorshDeserialize;
 
     /// Return the state commitment.
     fn state_commitment(&self) -> Self::StateCommitment;