--- a/grug/ffi/src/exports.rs
+++ b/grug/ffi/src/exports.rs
@@ -3,7 +3,7 @@ use {
     grug_types::{
         make_auth_ctx, make_immutable_ctx, make_mutable_ctx, make_sudo_ctx,
         unwrap_into_generic_result, AuthCtx, AuthResponse, BankMsg, BankQuery, BankQueryResponse,
-        BorshDeExt, Context, GenericResult, ImmutableCtx, Json, JsonDeExt, JsonSerExt, MutableCtx,
+        BorshDeExt, BorshSerExt, Context, GenericResult, ImmutableCtx, Json, JsonDeExt, MutableCtx,
         QuerierWrapper, Response, SubMsgResult, SudoCtx, Tx, TxOutcome,
     },
     serde::de::DeserializeOwned,
@@ -43,12 +43,14 @@ where
     let res = (|| {
         let ctx: Context = unwrap_into_generic_result!(ctx_bytes.deserialize_borsh());
         let ctx = make_mutable_ctx!(ctx, &mut ExternalStorage, &ExternalApi, &ExternalQuerier);
-        let msg = unwrap_into_generic_result!(msg_bytes.deserialize_json());
+
+        let msg = unwrap_into_generic_result!(msg_bytes.deserialize_borsh::<Json>());
+        let msg = unwrap_into_generic_result!(msg.deserialize_json());
 
         instantiate_fn(ctx, msg).into()
     })();
 
-    let res_bytes = res.to_json_vec().unwrap();
+    let res_bytes = res.to_borsh_vec().unwrap();
 
     Region::release_buffer(res_bytes) as usize
 }
@@ -68,12 +70,14 @@ where
     let res = (|| {
         let ctx: Context = unwrap_into_generic_result!(ctx_bytes.deserialize_borsh());
         let ctx = make_mutable_ctx!(ctx, &mut ExternalStorage, &ExternalApi, &ExternalQuerier);
-        let msg = unwrap_into_generic_result!(msg_bytes.deserialize_json());
+
+        let msg = unwrap_into_generic_result!(msg_bytes.deserialize_borsh::<Json>());
+        let msg = unwrap_into_generic_result!(msg.deserialize_json());
 
         execute_fn(ctx, msg).into()
     })();
 
-    let res_bytes = res.to_json_vec().unwrap();
+    let res_bytes = res.to_borsh_vec().unwrap();
 
     Region::release_buffer(res_bytes) as usize
 }
@@ -94,11 +98,13 @@ where
         let ctx: Context = unwrap_into_generic_result!(ctx_bytes.deserialize_borsh());
         let immutable_ctx =
             make_immutable_ctx!(ctx, &ExternalStorage, &ExternalApi, &ExternalQuerier);
-        let msg = unwrap_into_generic_result!(msg_bytes.deserialize_json());
+
+        let msg = unwrap_into_generic_result!(msg_bytes.deserialize_borsh::<Json>());
+        let msg = unwrap_into_generic_result!(msg.deserialize_json());
 
         query_fn(immutable_ctx, msg).into()
     })();
-    let res_bytes = res.to_json_vec().unwrap();
+    let res_bytes = res.to_borsh_vec().unwrap();
 
     Region::release_buffer(res_bytes) as usize
 }
@@ -118,12 +124,14 @@ where
     let res = (|| {
         let ctx: Context = unwrap_into_generic_result!(ctx_bytes.deserialize_borsh());
         let ctx = make_mutable_ctx!(ctx, &mut ExternalStorage, &ExternalApi, &ExternalQuerier);
-        let msg = unwrap_into_generic_result!(msg_bytes.deserialize_json());
+
+        let msg = unwrap_into_generic_result!(msg_bytes.deserialize_borsh::<Json>());
+        let msg = unwrap_into_generic_result!(msg.deserialize_json());
 
         migrate_fn(ctx, msg).into()
     })();
 
-    let res_bytes = res.to_json_vec().unwrap();
+    let res_bytes = res.to_borsh_vec().unwrap();
 
     Region::release_buffer(res_bytes) as usize
 }
@@ -145,13 +153,16 @@ where
     let res = (|| {
         let ctx: Context = unwrap_into_generic_result!(ctx_bytes.deserialize_borsh());
         let ctx = make_sudo_ctx!(ctx, &mut ExternalStorage, &ExternalApi, &ExternalQuerier);
-        let msg = unwrap_into_generic_result!(msg_bytes.deserialize_json());
-        let events = unwrap_into_generic_result!(events_bytes.deserialize_json());
+
+        let msg = unwrap_into_generic_result!(msg_bytes.deserialize_borsh::<Json>());
+        let msg = unwrap_into_generic_result!(msg.deserialize_json());
+
+        let events = unwrap_into_generic_result!(events_bytes.deserialize_borsh());
 
         reply_fn(ctx, msg, events).into()
     })();
 
-    let res_bytes = res.to_json_vec().unwrap();
+    let res_bytes = res.to_borsh_vec().unwrap();
 
     Region::release_buffer(res_bytes) as usize
 }
@@ -172,7 +183,7 @@ where
         receive_fn(ctx).into()
     })();
 
-    let res_bytes = res.to_json_vec().unwrap();
+    let res_bytes = res.to_borsh_vec().unwrap();
 
     Region::release_buffer(res_bytes) as usize
 }
@@ -193,7 +204,7 @@ where
         cron_execute_fn(ctx).into()
     })();
 
-    let res_bytes = res.to_json_vec().unwrap();
+    let res_bytes = res.to_borsh_vec().unwrap();
 
     Region::release_buffer(res_bytes) as usize
 }
@@ -212,12 +223,12 @@ where
     let res = (|| {
         let ctx: Context = unwrap_into_generic_result!(ctx_bytes.deserialize_borsh());
         let ctx = make_auth_ctx!(ctx, &mut ExternalStorage, &ExternalApi, &ExternalQuerier);
-        let tx = unwrap_into_generic_result!(tx_bytes.deserialize_json());
+        let tx = unwrap_into_generic_result!(tx_bytes.deserialize_borsh());
 
         authenticate_fn(ctx, tx).into()
     })();
 
-    let res_bytes = res.to_json_vec().unwrap();
+    let res_bytes = res.to_borsh_vec().unwrap();
 
     Region::release_buffer(res_bytes) as usize
 }
@@ -236,12 +247,12 @@ where
     let res = (|| {
         let ctx: Context = unwrap_into_generic_result!(ctx_bytes.deserialize_borsh());
         let ctx = make_auth_ctx!(ctx, &mut ExternalStorage, &ExternalApi, &ExternalQuerier);
-        let tx = unwrap_into_generic_result!(tx_bytes.deserialize_json());
+        let tx = unwrap_into_generic_result!(tx_bytes.deserialize_borsh());
 
         backrun_fn(ctx, tx).into()
     })();
 
-    let res_bytes = res.to_json_vec().unwrap();
+    let res_bytes = res.to_borsh_vec().unwrap();
 
     Region::release_buffer(res_bytes) as usize
 }
@@ -260,12 +271,12 @@ where
     let res = (|| {
         let ctx: Context = unwrap_into_generic_result!(ctx_bytes.deserialize_borsh());
         let ctx = make_sudo_ctx!(ctx, &mut ExternalStorage, &ExternalApi, &ExternalQuerier);
-        let msg = unwrap_into_generic_result!(msg_bytes.deserialize_json());
+        let msg = unwrap_into_generic_result!(msg_bytes.deserialize_borsh());
 
         transfer_fn(ctx, msg).into()
     })();
 
-    let res_bytes = res.to_json_vec().unwrap();
+    let res_bytes = res.to_borsh_vec().unwrap();
 
     Region::release_buffer(res_bytes) as usize
 }
@@ -284,12 +295,12 @@ where
     let res = (|| {
         let ctx: Context = unwrap_into_generic_result!(ctx_bytes.deserialize_borsh());
         let ctx = make_immutable_ctx!(ctx, &ExternalStorage, &ExternalApi, &ExternalQuerier);
-        let msg = unwrap_into_generic_result!(msg_bytes.deserialize_json());
+        let msg = unwrap_into_generic_result!(msg_bytes.deserialize_borsh());
 
         query_fn(ctx, msg).into()
     })();
 
-    let res_bytes = res.to_json_vec().unwrap();
+    let res_bytes = res.to_borsh_vec().unwrap();
 
     Region::release_buffer(res_bytes) as usize
 }
@@ -308,12 +319,12 @@ where
     let res = (|| {
         let ctx: Context = unwrap_into_generic_result!(ctx_bytes.deserialize_borsh());
         let auth_ctx = make_auth_ctx!(ctx, &mut ExternalStorage, &ExternalApi, &ExternalQuerier);
-        let tx = unwrap_into_generic_result!(tx_bytes.deserialize_json());
+        let tx = unwrap_into_generic_result!(tx_bytes.deserialize_borsh());
 
         withhold_fee_fn(auth_ctx, tx).into()
     })();
 
-    let res_bytes = res.to_json_vec().unwrap();
+    let res_bytes = res.to_borsh_vec().unwrap();
 
     Region::release_buffer(res_bytes) as usize
 }
@@ -334,13 +345,13 @@ where
     let res = (|| {
         let ctx: Context = unwrap_into_generic_result!(ctx_bytes.deserialize_borsh());
         let auth_ctx = make_auth_ctx!(ctx, &mut ExternalStorage, &ExternalApi, &ExternalQuerier);
-        let tx = unwrap_into_generic_result!(tx_bytes.deserialize_json());
-        let outcome = unwrap_into_generic_result!(outcome_bytes.deserialize_json());
+        let tx = unwrap_into_generic_result!(tx_bytes.deserialize_borsh());
+        let outcome = unwrap_into_generic_result!(outcome_bytes.deserialize_borsh());
 
         finalize_fee_fn(auth_ctx, tx, outcome).into()
     })();
 
-    let res_bytes = res.to_json_vec().unwrap();
+    let res_bytes = res.to_borsh_vec().unwrap();
 
     Region::release_buffer(res_bytes) as usize
 }