Reduction in contract sizes:

| contract              | before (kiB) | after (KiB) | change |
| --------------------- | ------------ | ----------- | ------ |
| dango-account-factory | 787          | 683         | -13%   |
| dango-account-safe    | 742          | 766         | +3%    |
| dango-account-spot    | 560          | 496         | -11%   |
| dango-amm             | 552          | 473         | -14%   |
| dango-bank            | 523          | 427         | -18%   |
| dango-ibc-transfer    | 366          | 269         | -26%   |
| dango-taxman          | 538          | 416         | -23%   |
| dango-token-factory   | 459          | 380         | -17%   |
| grug-mock-account     | 493          | 447         | -9%    |
| grug-mock-bank        | 510          | 394         | -23%   |
| grug-mock-taxman      | 536          | 414         | -23%   |
| grug-tester           | 406          | 368         | -9%    |

Improvement in TPS:

| benchmark | before (TPS) | after (TPS) | change |
| --------- | ------------ | ----------- | ------ |
| sends     | 1,712        | 1,962       | +15%   |
| swaps     | 648          | 764         | +18%   |